Moved bnd Configuration load to reusable Helper Class
