+++ biz.aQute.bnd.maven/src/aQute/bnd/maven/lib/configuration/package-info.java
@@ -1,4 +1,4 @@
-@Version("1.1.0")
+@Version("1.2.0")
 @Export
 package aQute.bnd.maven.lib.configuration;
